Adding stages and some other checks to the mega script

+++ SLURM/scripts/mega.sh
@@ -49,7 +49,7 @@ then
     load_gpu="module load gcccuda/2016a;module load CUDA/8.0.54;" 
     # Juicer directory, contains scripts/, references/, and restriction_sites/
     # can also be set in options via -D
-    juiceDir="/work/ea14/juicer" ### RICE
+    juiceDir="/projects/ea14/juicer" ### RICE
     # default queue, can also be set in options via -q
     queue="commons"
     # default long queue, can also be set in options via -l
@@ -98,6 +98,7 @@ usageHelp="Usage: ${0##*/} -g genomeID [-d topDir] [-s site] [-h]"
 genomeHelp="   genomeID must be defined in the script, e.g. \"hg19\" or \"mm10\" (default \"$genomeID\")"
 dirHelp="   [topDir] is the top level directory (default \"$topDir\") and must contain links to all merged_nodups files underneath it"
 siteHelp="   [site] must be defined in the script, e.g.  \"HindIII\" or \"MboI\" (default \"$site\"); alternatively, this can be the restriction site file"
+stageHelp="* [stage]: must be one of \"final\", \"postproc\", or \"early\".\n    -Use \"final\" when the reads have been combined into merged_nodups but the\n     final stats and hic files have not yet been created.\n    -Use \"postproc\" when the hic files have been created and only\n     postprocessing feature annotation remains to be completed.\n    -Use \"early\" for an early exit, before the final creation of the stats and\n     hic files"
 excludeHelp="   -x: exclude fragment-delimited maps from Hi-C mega map (will run much faster)"
 helpHelp="   -h: print this help and exit"
 
@@ -106,32 +107,34 @@ printHelpAndExit() {
     echo "$genomeHelp"
     echo "$dirHelp"
     echo "$siteHelp"
+    echo "$stageHelp"
     echo "$excludeHelp"
     echo "$helpHelp"
     exit "$1"
 }
 
-while getopts "d:g:hxs:" opt; do
+while getopts "d:g:hxs:S:" opt; do
     case $opt in
 	g) genomeID=$OPTARG ;;
 	h) printHelpAndExit 0;;
 	d) topDir=$OPTARG ;;
 	s) site=$OPTARG ;;
-  x) exclude=1 ;;
+	x) exclude=1 ;;
+	S) stage=$OPTARG ;;
 	[?]) printHelpAndExit 1;;
     esac
 done
 
 ## Set ligation junction based on restriction enzyme
 case $site in
-	HindIII) ligation="AAGCTAGCTT";;
-	DpnII) ligation="GATCGATC";;
-	MboI) ligation="GATCGATC";;
+    HindIII) ligation="AAGCTAGCTT";;
+    DpnII) ligation="GATCGATC";;
+    MboI) ligation="GATCGATC";;
     none) ligation="XXXX";;
-	*)  ligation="XXXX"
-      site_file=$site
-      echo "$site not listed as recognized enzyme, so trying it as site file."
-      echo "Ligation junction is undefined";;
+    *)  ligation="XXXX"
+	site_file=$site
+	echo "$site not listed as recognized enzyme, so trying it as site file."
+	echo "Ligation junction is undefined";;
 esac
 
 if [ -z "$site_file" ]
@@ -144,7 +147,19 @@ if [ ! -e "$site_file" ] && [ "$site" != "none" ]
 then
     echo "***! $site_file does not exist. It must be created before running this script."
     echo "The site file is used for statistics even if fragment delimited maps are excluded"
-    exit 100
+    exit 1
+fi
+
+if [ ! -z "$stage" ]
+then
+    case $stage in
+        final) final=1 ;;
+	early) early=1 ;;
+	postproc) postproc=1 ;;
+        *)  echo "$usageHelp"
+	    echo "$stageHelp"
+	    exit 1
+    esac
 fi
 
 ## Directories to be created and regex strings for listing files
@@ -156,22 +171,32 @@ outfile=${megadir}/lsf.out
 #output messages
 logdir="$megadir/debug"
 touchfile1=${megadir}/touch1
+touchfile2=${megadir}/touch2
+touchfile3=${megadir}/touch3
+touchfile4=${megadir}/touch4
+touchfile5=${megadir}/touch5
+touchfile6=${megadir}/touch6
+touchfile7=${megadir}/touch7
+touchfile8=${megadir}/touch8
 
-
-## Check for existing merge_nodups files:
+## Check for existing merged_nodups files:
 
 merged_count=`find -L ${topDir} | grep merged_nodups.txt | wc -l`
 if [ "$merged_count" -lt "1" ]
 then
 	echo "***! Failed to find at least one merged_nodups files under ${topDir}"
-	exit 100
+	exit 1
 fi
 
-merged_names=$(find -L ${topDir} | grep merged_nodups.txt | tr '\n' ' ')
+merged_names=$(find -L ${topDir} | grep merged_nodups.txt.gz | awk '{print "<(gunzip -c",$1")"}' | tr '\n' ' ')
+if [ ${#merged_names} -eq 0 ]
+then
+    merged_names=$(find -L ${topDir} | grep merged_nodups.txt | tr '\n' ' ')
+fi
 inter_names=$(find -L ${topDir} | grep inter.txt | tr '\n' ' ')
 
 ## Create output directory, exit if already exists
-if [[ -d "${outputdir}" ]] 
+if [[ -d "${outputdir}" ]] && [ -z $final ] && [ -z $postproc ]
 then
     echo "***! Move or remove directory \"${outputdir}\" before proceeding."
 	exit 101			
@@ -193,12 +218,13 @@ fi
 
 ## Arguments have been checked and directories created. Now begins
 ## the real work of the pipeline
-#source $usePath
-#$load_cluster
 
+# Not in final or postproc
+if [ -z $final ] && [ -z $postproc ]
+then
 # Create top statistics file from all inter.txt files found under current dir
-touchfile1=${megadir}/touch1
-jid1=`sbatch <<- TOPSTATS | egrep -o -e "\b[0-9]+$"
+
+    jid1=`sbatch <<- TOPSTATS | egrep -o -e "\b[0-9]+$"
 #!/bin/bash -l
 #SBATCH -p $queue
 #SBATCH -t 1440
@@ -207,6 +233,7 @@ jid1=`sbatch <<- TOPSTATS | egrep -o -e "\b[0-9]+$"
 #SBATCH -o $logdir/topstats-%j.out
 #SBATCH -e $logdir/topstats-%j.err
 #SBATCH -J "${groupname}_topstats"
+#SBATCH --mem-per-cpu=32G 
 export LC_COLLATE=C
 if ! awk -f ${juiceDir}/scripts/makemega_addstats.awk ${inter_names} > ${outputdir}/inter.txt
 then  
@@ -218,11 +245,10 @@ else
 fi
 touch $touchfile1
 TOPSTATS`
-dependtopstats="afterok:$jid1"
+    dependtopstats="afterok:$jid1"
 
 # Merge all merged_nodups.txt files found under current dir
-touchfile2=${megadir}/touch2
-jid2=`sbatch <<- MRGSRT | egrep -o -e "\b[0-9]+$"
+    jid2=`sbatch <<- MRGSRT | egrep -o -e "\b[0-9]+$"
 #!/bin/bash -l
 #SBATCH -p ${long_queue}
 #SBATCH -t ${long_queue_time}
@@ -233,26 +259,42 @@ jid2=`sbatch <<- MRGSRT | egrep -o -e "\b[0-9]+$"
 #SBATCH -e $logdir/merge-%j.err
 #SBATCH -J "${groupname}_merge"
 #SBATCH -d "${dependtopstats}"
+#SBATCH --mem-per-cpu=32G 
 if [ ! -f "${touchfile1}" ]
 then
     echo "***! Top stats job failed, type \"scontrol show job $jid1\" to see what happened."
     exit 1
 fi
-if ! sort -T ${tmpdir} -m -k2,2d -k6,6d ${merged_names} > ${outputdir}/merged_nodups.txt
-then 
-    echo "***! Some problems occurred somewhere in creating sorted merged_nodups files."
-    exit 100
-else
-    echo "(-: Finished sorting all merged_nodups files into a single merge."
-    rm -r ${tmpdir}
-fi
+
+  if [ $isRice -eq 1 ]
+  then
+    if ! /poscratch/aidenlab/olga/sort --parallel=48 -S8G -T ${tmpdir} -m -k2,2d -k6,6d ${merged_names} > ${outputdir}/merged_nodups.txt
+    then
+      echo "***! Some problems occurred somewhere in creating sorted merged_nodups files."
+      exit 1
+    fi
+  else
+    if ! sort -T ${tmpdir} -m -k2,2d -k6,6d ${merged_names} > ${outputdir}/merged_nodups.txt
+    then 
+      echo "***! Some problems occurred somewhere in creating sorted merged_nodups files."
+      exit 1
+    else
+      echo "(-: Finished sorting all merged_nodups files into a single merge."
+      rm -r ${tmpdir}
+    fi
+  fi
 touch $touchfile2
 MRGSRT`
-dependmerge="afterok:$jid2"
+    dependmerge="#SBATCH -d afterok:$jid2"
+else
+    touch $touchfile1
+    touch $touchfile2
+fi
 
-# Create statistics files for MQ > 0
-touchfile3=${megadir}/touch3  
-jid3=`sbatch <<- INTER0 | egrep -o -e "\b[0-9]+$"
+if [ -z $postproc ] && [ -z $early ]
+then
+    # Create statistics files for MQ > 0
+    jid3=`sbatch <<- INTER0 | egrep -o -e "\b[0-9]+$"
 #!/bin/bash -l
 #SBATCH -p ${long_queue}
 #SBATCH -t ${long_queue_time}
@@ -261,20 +303,22 @@ jid3=`sbatch <<- INTER0 | egrep -o -e "\b[0-9]+$"
 #SBATCH -o $logdir/inter0-%j.out
 #SBATCH -e $logdir/inter0-%j.err
 #SBATCH -J "${groupname}_inter0"
-#SBATCH -d "${dependmerge}"
+#SBATCH --mem-per-cpu=32G 
+${dependmerge}
 if [ ! -f "${touchfile2}" ]
 then
    echo "***! Sort job failed."
    exit 1
 fi
-${juiceDir}/scripts/statistics.pl -q 1 -o${outputdir}/inter.txt -s $site_file -l $ligation ${outputdir}/merged_nodups.txt
-touch $touchfile3
+if ${juiceDir}/scripts/statistics.pl -q 1 -o${outputdir}/inter.txt -s $site_file -l $ligation ${outputdir}/merged_nodups.txt
+then
+   touch $touchfile3
+fi
 INTER0`
-dependinter0="afterok:$jid3"
+    dependinter0="afterok:$jid3"
 
-# Create statistics files for MQ > 30
-touchfile4=${megadir}/touch4
-jid4=`sbatch <<- INTER30 | egrep -o -e "\b[0-9]+$"
+    # Create statistics files for MQ > 30
+    jid4=`sbatch <<- INTER30 | egrep -o -e "\b[0-9]+$"
 #!/bin/bash -l
 #SBATCH -p ${long_queue}
 #SBATCH -t ${long_queue_time}
@@ -283,20 +327,22 @@ jid4=`sbatch <<- INTER30 | egrep -o -e "\b[0-9]+$"
 #SBATCH -o $logdir/inter30-%j.out
 #SBATCH -e $logdir/inter30-%j.err
 #SBATCH -J "${groupname}_inter30"
-#SBATCH -d "${dependmerge}"
+#SBATCH --mem-per-cpu=32G 
+${dependmerge}
 if [ ! -f "${touchfile2}" ]
 then
    echo "***! Sort job failed."
    exit 1
 fi
-${juiceDir}/scripts/statistics.pl -q 30 -o${outputdir}/inter_30.txt -s $site_file -l $ligation ${outputdir}/merged_nodups.txt 
-touch $touchfile4
+if ${juiceDir}/scripts/statistics.pl -q 30 -o${outputdir}/inter_30.txt -s $site_file -l $ligation ${outputdir}/merged_nodups.txt 
+then
+  touch $touchfile4
+fi
 INTER30`
-dependinter30="afterok:$jid4"
+    dependinter30="afterok:$jid4"
 
-# Create HIC maps file for MQ > 0
-touchfile5=${megadir}/touch5
-jid5=`sbatch <<- HIC0 | egrep -o -e "\b[0-9]+$"
+    # Create HIC maps file for MQ > 0
+    jid5=`sbatch <<- HIC0 | egrep -o -e "\b[0-9]+$"
 #!/bin/bash -l
 #SBATCH -p ${long_queue}
 #SBATCH -t ${long_queue_time}
@@ -322,11 +368,10 @@ then
    touch $touchfile5 
 fi
 HIC0`
-dependhic0="afterok:$jid5"
+    dependhic0="afterok:$jid5"
 
-# Create HIC maps file for MQ > 30
-touchfile6=${megadir}/touch6
-jid6=`sbatch <<- HIC30 | egrep -o -e "\b[0-9]+$"
+    # Create HIC maps file for MQ > 30
+    jid6=`sbatch <<- HIC30 | egrep -o -e "\b[0-9]+$"
 #!/bin/bash -l
 #SBATCH -p ${long_queue}
 #SBATCH -t ${long_queue_time}
@@ -352,18 +397,25 @@ then
    touch $touchfile6
 fi
 HIC30`
-dependhic30="${dependhic0}:$jid6"
-dependhic30only="afterok:$jid6"
+    dependhic30only="afterok:$jid6"
+    sbatchdepend="#SBATCH -d ${dependhic30only}"
+    dependhic30="${dependhic0}:$jid6"
+else
+    touch $touchfile3 $touchfile4 $touchfile5 $touchfile6
+    sbatchdepend=""
+fi
 
-# Create loop lists file for MQ > 30
-touchfile7=${megadir}/touch7
-if [ $isRice -eq 1 ] || [ $isVoltron -eq 1 ]
+if [ -z $early ]
 then
-    if [  $isRice -eq 1 ]
+# Create loop lists file for MQ > 30
+    touchfile7=${megadir}/touch7
+    if [ $isRice -eq 1 ] || [ $isVoltron -eq 1 ]
     then
-	sbatch_req="#SBATCH --gres=gpu:kepler:1"
-    fi
-    jid7=`sbatch <<- HICCUPS | egrep -o -e "\b[0-9]+$"
+	if [  $isRice -eq 1 ]
+	then
+	    sbatch_req="#SBATCH --gres=gpu:kepler:1"
+	fi
+	jid7=`sbatch <<- HICCUPS | egrep -o -e "\b[0-9]+$"
 	#!/bin/bash -l
 	#SBATCH -p ${long_queue}
 	#SBATCH -t 1440
@@ -373,7 +425,7 @@ then
 	#SBATCH -o $logdir/hiccups-%j.out
 	#SBATCH -e $logdir/hiccups-%j.err
 	#SBATCH -J "${groupname}_hiccups"
-	#SBATCH -d "${dependhic30only}"
+	${sbatchdepend}
 	${sbatch_req}
 	$load_java
 	if [ ! -f "${touchfile6}" ]
@@ -385,15 +437,14 @@ then
 	${juiceDir}/scripts/juicer_hiccups.sh -j ${juiceDir}/scripts/juicer_tools -i $outputdir/inter_30.hic -m ${juiceDir}/references/motif -g $genomeID
 	touch $touchfile7
 	HICCUPS`
-    dependhiccups="afterok:$jid7"
-else
-    dependhiccups="afterok"
-    touch $touchfile7
-fi
+	dependhic30="${dependhic30}:$jid7"
+    else
+	touch $touchfile7
+    fi
 
-touchfile8=${megadir}/touch8
-# Create domain lists for MQ > 30
-jid8=`sbatch <<- ARROWHEAD | egrep -o -e "\b[0-9]+$"
+    touchfile8=${megadir}/touch8
+    # Create domain lists for MQ > 30
+    jid8=`sbatch <<- ARROWHEAD | egrep -o -e "\b[0-9]+$"
 #!/bin/bash -l
 #SBATCH -p ${long_queue}
 #SBATCH -t ${long_queue_time}
@@ -403,8 +454,7 @@ jid8=`sbatch <<- ARROWHEAD | egrep -o -e "\b[0-9]+$"
 #SBATCH -o $logdir/arrowhead-%j.out
 #SBATCH -e $logdir/arrowhead-%j.err
 #SBATCH -J "${groupname}_arrowhead"
-#SBATCH -d "${dependhic30only}"
-#source $usePath
+${sbatchdepend}
 $load_java	
 if [ ! -f "${touchfile6}" ]
 then
@@ -414,19 +464,18 @@ fi
 ${juiceDir}/scripts/juicer_arrowhead.sh -j ${juiceDir}/scripts/juicer_tools -i $outputdir/inter_30.hic
 touch $touchfile8
 ARROWHEAD`
-dependarrows="${dependhiccups}:$jid8"
-
-# Final checks
-jid9=`sbatch <<- FINAL | egrep -o -e "\b[0-9]+$"
+    dependhic30="${dependhic0}:$jid8"
+    # Final checks
+    jid9=`sbatch <<- FINAL | egrep -o -e "\b[0-9]+$"
 #!/bin/bash -l
-#SBATCH -p ${long_queue}
+#SBATCH -p ${queue}
 #SBATCH -t 100
 #SBATCH -c 1
 #SBATCH --ntasks=1
 #SBATCH -o $logdir/done-%j.out
 #SBATCH -e $logdir/done-%j.err
 #SBATCH -J "${groupname}_done"
-#SBATCH -d "${dependarrows}"
+#SBATCH -d "${dependhic30}"
 
 if [ ! -f "${touchfile5}" ]
 then
@@ -456,4 +505,21 @@ then
    echo $topDir, $site, $genomeID, $genomePath | mail -r MegaJuicer@rice.edu -s \"Mega Juicer pipeline finished successfully @ Rice\" -t $USER@rice.edu;
 fi
 FINAL`
+else
+    jid9=`sbatch <<- FINAL | egrep -o -e "\b[0-9]+$"
+#!/bin/bash -l
+#SBATCH -p ${queue}
+#SBATCH -t 100
+#SBATCH -c 1
+#SBATCH --ntasks=1
+#SBATCH -o $logdir/done-%j.out
+#SBATCH -e $logdir/done-%j.err
+#SBATCH -J "${groupname}_done"
+#SBATCH -d "${dependmerge}"
+
+rm -r ${tmpdir}
+rm -f $touchfile1 $touchfile2 $touchfile3 $touchfile4 $touchfile5 $touchfile6 $touchfile7 $touchfile8
+echo "(-: Successfully completed making mega map. Done. :-)"
+FINAL`
+fi
 echo "(-: Finished adding all jobs... please wait while processing."
